fixed tests to work with new routing/layout
