+++ interface/__tests__/components/UseCaseModal.test.tsx
@@ -1,5 +1,11 @@
 import React from "react";
-import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+import {
+  act,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from "@testing-library/react";
 import UseCaseModal from "../../components/UseCaseModal";
 import { Community } from "../../utils/account-requests";
 
@@ -10,24 +16,30 @@ jest.mock("react-router-dom", () => ({
 
 const existingScorers = [
   {
-    "name": "Existing 1",
-    "description": "asdfasdf",
-    "id": 7,
-    "created_at": "2023-03-20T20:27:02.425Z",
-    "use_case": "Airdrop Protection"
+    name: "Existing 1",
+    description: "asdfasdf",
+    id: 7,
+    created_at: "2023-03-20T20:27:02.425Z",
+    use_case: "Airdrop Protection",
   },
   {
-    "name": "Existing 2",
-    "description": "a",
-    "id": 8,
-    "created_at": "2023-03-20T22:19:22.363Z",
-    "use_case": "Airdrop Protection"
-  }
+    name: "Existing 2",
+    description: "a",
+    id: 8,
+    created_at: "2023-03-20T22:19:22.363Z",
+    use_case: "Airdrop Protection",
+  },
 ] as Community[];
 
 describe("UseCaseModal", () => {
   it("should display a list of use cases", async () => {
-    render(<UseCaseModal existingScorers={existingScorers} isOpen={true} onClose={() => { }} />);
+    render(
+      <UseCaseModal
+        existingScorers={existingScorers}
+        isOpen={true}
+        onClose={() => {}}
+      />
+    );
     const useCaseItems = screen.getAllByTestId("use-case-item");
 
     expect(screen.getByText("Select a Use Case")).toBeInTheDocument();
@@ -35,7 +47,13 @@ describe("UseCaseModal", () => {
   });
 
   it("continue button should only be enabled when a use case is selected", async () => {
-    render(<UseCaseModal existingScorers={existingScorers} isOpen={true} onClose={() => { }} />);
+    render(
+      <UseCaseModal
+        existingScorers={existingScorers}
+        isOpen={true}
+        onClose={() => {}}
+      />
+    );
 
     const useCaseItem = screen.getAllByTestId("use-case-item")[0];
     const continueButton = screen.getByText(/Continue/i).closest("button");
@@ -48,7 +66,13 @@ describe("UseCaseModal", () => {
   });
 
   it("should switch to use case details step when continue button is clicked on first step", async () => {
-    render(<UseCaseModal existingScorers={existingScorers} isOpen={true} onClose={() => { }} />);
+    render(
+      <UseCaseModal
+        existingScorers={existingScorers}
+        isOpen={true}
+        onClose={() => {}}
+      />
+    );
 
     const useCaseItem = screen.getAllByTestId("use-case-item")[0];
     const continueButton = screen.getByText(/Continue/i).closest("button");
@@ -61,7 +85,13 @@ describe("UseCaseModal", () => {
   });
 
   it("continue button should only be enabled when a use case name and description is filled", async () => {
-    render(<UseCaseModal existingScorers={existingScorers} isOpen={true} onClose={() => { }} />);
+    render(
+      <UseCaseModal
+        existingScorers={existingScorers}
+        isOpen={true}
+        onClose={() => {}}
+      />
+    );
 
     const useCaseItem = screen.getAllByTestId("use-case-item")[0];
     const firstContinueButton = screen.getByText(/Continue/i).closest("button");
@@ -87,7 +117,13 @@ describe("UseCaseModal", () => {
   });
 
   it("should show duplication error if scorer has duplicate name", async () => {
-    render(<UseCaseModal existingScorers={existingScorers} isOpen={true} onClose={() => { }} />);
+    render(
+      <UseCaseModal
+        existingScorers={existingScorers}
+        isOpen={true}
+        onClose={() => {}}
+      />
+    );
     const useCaseItem = screen.getAllByTestId("use-case-item")[0];
     const firstContinueButton = screen.getByText(/Continue/i).closest("button");
 
@@ -108,10 +144,15 @@ describe("UseCaseModal", () => {
       target: { value: "This is my use case description" },
     });
 
-    await secondContinueButton?.click();
+    expect(secondContinueButton).toBeEnabled();
+
+    if (!secondContinueButton) throw new Error("Button not found");
+    fireEvent.click(secondContinueButton);
 
     await waitFor(() => {
-      expect(screen.getByText("A scorer with this name already exists")).toBeInTheDocument();
+      expect(
+        screen.getByText("A scorer with this name already exists")
+      ).toBeInTheDocument();
     });
   });
 });
